+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.2.9.1
 
 -   Increase width of dropdown when selecting a curse mod file
+-   Add some basic filtering to not show files for a different loader